+++ src/components/Results/ResultTab.vue
@@ -1,15 +1,4 @@
 <template>
-  <div>
-    <router-link to="/results/200">
-      <Icon name="arrow-left" />
-    </router-link>
-    <div>
-      <div v-for="(tab, index) in detailTabs" :key="index">
-        <router-link :to="getTabLink(tab)">{{ tab }}</router-link>
-      </div>
-    </div>
-  </div>
-  <!-- <information-summary :information="summaryProps"></information-summary> -->
   <Individual v-if="currentTabComponent === 'individual'" />
   <Statistics v-if="currentTabComponent === 'statistics'" />
   <Spreadsheet v-if="currentTabComponent === 'spreadsheet'" />
@@ -21,7 +10,6 @@ import { useRoute } from 'vue-router'
 import Individual from './Individual.vue'
 import Statistics from './Statistics.vue'
 import Spreadsheet from './Spreadsheet.vue'
-import Icon from '/@/components/UI/Icon.vue'
 import { detailTabs } from './use/utils'
 
 export default defineComponent({
@@ -29,8 +17,7 @@ export default defineComponent({
   components: {
     Individual,
     Statistics,
-    Spreadsheet,
-    Icon
+    Spreadsheet
   },
   setup() {
     const selectedTab = ref('Statistics')
@@ -44,15 +31,6 @@ export default defineComponent({
       return ''
     })
 
-    // link用のoption
-    const getTabLink = (tab: string) => ({
-      name: 'results',
-      params: { id: route.params.id },
-      query: {
-        tab: tab
-      }
-    })
-
     watch(
       () => route.query,
       async newQuery => {
@@ -61,9 +39,7 @@ export default defineComponent({
     )
     return {
       selectedTab,
-      currentTabComponent,
-      getTabLink,
-      detailTabs
+      currentTabComponent
     }
   }
 })
